+++ packages/urbanjs-tool-tslint/src/types.ts
@@ -2,6 +2,7 @@ export type TslintConfig = {
   files: string | string[];
   configFile: string;
   extensions: string[];
-  formatter: string;
-  rulesDirectory: string;
+  formatter?: string;
+  formattersDirectory?: string;
+  rulesDirectory?: string;
 };