+++ packages/urbanjs-tool-webpack/tests/valid-project/gulpfile.js
@@ -1,12 +1,13 @@
 'use strict';
 
 const gulp = require('gulp');
+const path = require('path');
 const tools = require('@tamasmagedli/urbanjs-tools');
 
 tools.tasks.webpack.register(gulp, 'webpack', {
   entry: './index.js',
   output: {
-    path: './dist',
+    path: path.join(process.cwd(), './dist'),
     filename: 'index.js',
     libraryTarget: 'commonjs'
   },